Move polling related tests to a new file
